Refactor deliver grant funding routes
Split them out into groups (mostly based around items in the service
navigation).
