+++ app/constants.py
@@ -1,1 +1,2 @@
 DEFAULT_SECTION_NAME = "Tasks"
+CHECK_YOUR_ANSWERS = "check-your-answers"
